fix(TorchCe): hide camera access notification on mobile devices
On mobile devices, a persistent notification is shown while the app is
accessing the camera (in order to use the torch). Since re-newing the
camera access introduces a noticeable lag, we don't want to have to do
it while the user is actively using the app.

When the app loses focus (e.g. switching to a different tab or app) and
the torch os off, it is safe to release the camera and re-new the access
as soon as the app is focused again. This way, the notification is only
shown while the app is focused or the torch is on.