cfg(linting): overhaul linting and switch from `tslint` to `eslint`
Since `tslint` has been discontinued, switch to `eslint` +
`@typescript/eslint` for linting TS files as well. This consisted of a
big enough overhaul of the linting process, that I piggy-backed on it to
make the following related changes as well:

- Add plugins/rules regarding imports.
- Add type-checking for the JS files in the project (e.g. scripts).
- Update output target to ES2022.
- Fix building, linting and type-checking issues, as well as tests.

As part of this work, I also updated all dependencies to latest versions
(with the exception of `eslint` being kept at v8.x, because some plugins
are not yet compatible with the latest v9.x version).