fix(TorchCe): rework `TorckCe` to work on modern devices
The previous implementation would only work if the first camera had a
torch. Given that most of modern devices have multiple cameras and only
one of them (usually the last one) has a torch, the implementation has
been reworked to be able to detect a torch on any of the cameras.