cfg(audit-web-app): use local `webhint` connector to work in headless environments
Although the [puppeteer][1] `webhint` connector supports the
`headless: true` option, it did not work (throwing an `AnalyzeError`).
Switch to the [local][2] `webhint` connector to support running audits
in headless environments (such as CI, GitHub Codepsaces, etc.).

As part of this migration, some of the audits had to be removed, since
the `local` connector does not support them.

[1]: https://webhint.io/docs/user-guide/connectors/connector-puppeteer
[2]: https://webhint.io/docs/user-guide/connectors/connector-local