cfg(eslint): improve `env` configuration
Given that `overrides` are all applied in order (if the glob pattern
matches), the browser environment should only be applied to files in
`src/app/` and not test `.ts` files.

Also, the `jasmine` environment was removed from test `.ts` files,
because (a) not all tests will use Jasmine and (b) it didn't seem to
have any effect (unfortunately).

> [!TIP]
> The `jasmine` ESLint plugin and rules were are still applied to all
> tests (even those that might not use Jasmine), because many of the
> rules still apply (with the use of `describe()` and `it()` functions).

[1]: https://eslint.org/docs/latest/use/configure/configuration-files-deprecated#configuration-based-on-glob-patterns