cfg(*): restructure the app files layout to better isolate app code from test code
Previously, the compiled test code was output in a directory alongside
the application code. This made it harder (and more error-prone) to
isolate the codes and performa operations on one of them only.

This commit improves the situation, by isolating the app code (in
`out/app/`) from the test code (in `out/test/`).

As part of this change, we also switch to using relative paths for all
resources that are loaded in the app (images, sounds). This brings
consistency across the app code and improves the portability of the app
(in case we want to "mount" it at a sub-path in the future).