fix(images): add transparency to the `simple-torch-180x180.png` image
The 180x180 image was accidentally generated by hand in commit
ebf10968f703404bcfae1d27077285b1bfc6b516. Add it to the list of image
sizes and generate it properly using the `generate-icons.mjs` script.

This also ensures that the image has a transparent background and is
consistent with the rest of the icons.