cfg(Firebase): fix Firebase local serving (`serve` or `emulators:start`)
Due to a [breaking change][1] in Node.js v22, `superstatic`, the local
server used by Firebase `serve` and `emulators:start` commands throws an
error. See, also, [firebase/superstatic#468 (comment)][2] for details.

Apply a local patch to fix the issue, by accessing a property that is
defined on the `Stat` instance and thus will be available. We currently
do not need or use Firebase local serving, but I figured out the fix,
so... ¯\\\_(ツ)_/¯

[1]: https://github.com/nodejs/node/pull/50908
[2]: https://github.com/firebase/superstatic/issues/468#issuecomment-2166883064