fix(TorchCe): recover from invalid stored `torchDeviceId`
When a camera with torch is found, its device ID is stored in settings
for faster future retrieval. However, it turns out that device IDs can
sometimes change.

Previously, when the stored device ID was invalid (for example, because
the device ID had changed), the app was not able to recover, always
trying to retrieve the camera with the invalid device ID and throwing an
`OverconstrainedError` error.

This commit ensures that the stored device ID is only used if it
corresponds to an existing camera device ID, thus allowing the app to
recover from an invalid stored device ID.

> [!Note]
> This new implementation is slightly slower, because we have to
> enumerate the available media devices each time, but it is necessary
> for the app to work correctly when device IDs change.
> Alternatively, we could catch and ignore the `OverconstrainedError`
> error, but because browsers implement this differently, there is no
> reliable, cross-browser way to ensure we are only ignoring that
> particular error.